+++ src/masonite/commands/MakeMailableCommand.py
@@ -1,11 +1,11 @@
 """New Mailable Command."""
-from cleo import Command
 import inflection
 import os
 
 from ..utils.filesystem import make_directory, render_stub_file, get_module_dir
 from ..utils.str import as_filepath
 from ..utils.location import base_path
+from .Command import Command
 
 
 class MakeMailableCommand(Command):
@@ -14,6 +14,7 @@ class MakeMailableCommand(Command):
 
     mailable
         {name : Name of the mailable}
+        {--f|force=? : Force overriding file if already exists}
     """
 
     def __init__(self, application):
@@ -29,7 +30,11 @@ class MakeMailableCommand(Command):
         )
         filepath = base_path(relative_filename)
         make_directory(filepath)
-
+        if os.path.exists(filepath) and not self.option("force"):
+            self.warning(
+                f"{filepath} already exists! Run the command with -f (force) to override."
+            )
+            return -1
         with open(filepath, "w") as f:
             f.write(content)
