+++ src/masonite/commands/MakeProviderCommand.py
@@ -1,11 +1,11 @@
 """New Provider Command."""
-from cleo import Command
 import inflection
 import os
 
 from ..utils.filesystem import make_directory, render_stub_file, get_module_dir
 from ..utils.str import as_filepath
 from ..utils.location import base_path
+from .Command import Command
 
 
 class MakeProviderCommand(Command):
@@ -14,6 +14,7 @@ class MakeProviderCommand(Command):
 
     provider
         {name : Name of the provider}
+        {--f|force=? : Force overriding file if already exists}
     """
 
     def __init__(self, application):
@@ -30,6 +31,11 @@ class MakeProviderCommand(Command):
         )
         filepath = base_path(relative_filename)
         make_directory(filepath)
+        if os.path.exists(filepath) and not self.option("force"):
+            self.warning(
+                f"{filepath} already exists! Run the command with -f (force) to override."
+            )
+            return -1
 
         with open(filepath, "w") as f:
             f.write(content)
