+++ src/masonite/notification/commands/MakeNotificationCommand.py
@@ -1,11 +1,11 @@
 """New Notification Command"""
-from cleo import Command
 import inflection
 import os
 
 from ...utils.filesystem import get_module_dir, make_directory, render_stub_file
 from ...utils.location import base_path
 from ...utils.str import as_filepath
+from ...commands.Command import Command
 
 
 class MakeNotificationCommand(Command):
@@ -14,6 +14,7 @@ class MakeNotificationCommand(Command):
 
     notification
         {name : Name of the notification}
+        {--f|force=? : Force overriding file if already exists}
     """
 
     def __init__(self, application):
@@ -30,7 +31,11 @@ class MakeNotificationCommand(Command):
         )
         filepath = base_path(relative_filename)
         make_directory(filepath)
-
+        if os.path.exists(filepath) and not self.option("force"):
+            self.warning(
+                f"{filepath} already exists! Run the command with -f (force) to override."
+            )
+            return -1
         with open(filepath, "w") as f:
             f.write(content)
